requirements.txt
django-otp==1.5.4
qrcode==8.0
requests==2.32.3